fix a float32 bug

+++ cmd/web/render.go
@@ -29,7 +29,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	f := float32(n / 100)
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
